requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.2
anthropic==0.34.2
sentence-transformers==3.1.1
chromadb==0.5.15
rank-bm25==0.2.2
pdfplumber==0.11.4
ragas==0.1.21
langsmith==0.1.125
datasets==3.0.1
gradio==4.44.0
python-dotenv==1.0.1